% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/yknot.barchart.R
\name{yknot.barchart}
\alias{yknot.barchart}
\title{A Barchart function with ggplot2}
\usage{
yknot.barchart(df, x, y, fill, title, xlab, ylab, filename)
}
\arguments{
\item{df}{dataframe which x and y are from}

\item{x}{column of x}

\item{y}{column of y}

\item{fill}{column of fill (categorical data)}

\item{title}{title of plot}

\item{xlab}{label of x data}

\item{ylab}{label of y data}

\item{filename}{filename of the png}
}
\description{
Creates barchart using ggplot2 and saves it as png in current directory
}
\examples{
yknot.barchart(iris, "Sepal.Length", "Sepal.Width", "Species", "Sepal Comparison", "Sepal Length", "Sepal Width", "Test")
}
\keyword{plot}

