% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/yknot.barchart.R
\name{yknot.barchart}
\alias{yknot.barchart}
\title{A Barchart function with ggplot2}
\usage{
yknot.barchart(df, x, y, fill, title, xlab, ylab, filename)
}
\arguments{
\item{df}{dataframe
x x data
y ydata}
}
\description{
Creates barchart and saves it as png
}
\examples{
yknot.barchart(iris, "Sepal.Length", "Sepal.Width", "Species", "Sepal Comparison", "Sepal Length", "Sepal Width", "Test")
}
\keyword{plot}

